LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/inheritance/popup-inheritance-form-submission-expected.txt


FAIL A new frame's document.referrer is correct even if its initiator frame is gone assert_equals: expected "http://localhost:8800/referrer-policy/generic/inheritance/resources/submit-form-and-remove-frame.html" but got ""
